Hacky prototype to enable distributed development
Once we better understand what messages and objects need to be communicated between the client and workers we will want to replace this with a solution tailored specifically for the calls required to enable distributed Quake. This includes replacing the general RPC calls with more tightly focused ones as well as making the wrapper more transparent to users of regular Quake.